feat: add createGraphCavas, and use it in main.ts
