+++ packages/g6/__tests__/mock/index.ts
@@ -1,1 +1,2 @@
-export * from './graph';
+export { createGraph, createGraphCanvas } from './create';
+export { Graph } from './graph';
